Fixed app.hook.ts -> App.tsx
